fix(Enjoy2Live): refactored LabCampus & added urls
