+++ src/components/modules/UserProfile/LabCampus/LabCampus.js
@@ -1,186 +1,22 @@
-import React, { useEffect, useState } from "react";
-import {
-  Wrapper,
-  TimeLineContainer,
-  TimeStamp,
-  LabsContainer,
-  Lab,
-  AlignedDropdown,
-  Time,
-  Title
-} from "./StyledLabCampus";
+import React, { useState } from "react";
+import { Wrapper } from "./StyledLabCampus";
 let events = require("../../../../content/collections/timeline");
-import { ConvertCentralTime } from "../../../../utils/TimeZoneConverter";
-import ProgressBar from "@ramonak/react-progress-bar";
-import { DateTime, Interval } from "luxon";
+import { DateTime } from "luxon";
+import TimeLine from "./TimeLine";
+import Labs from "./Labs";
+
 const LabCampus = props => {
   const [currentTime, setCurrentTime] = useState(DateTime.now());
-
-  const handleProgression = eventInterval => {
-    let remainingLabMintues =
-      Interval.fromDateTimes(currentTime, eventInterval.end).count("minute") -
-      1;
-
-    const timeUpdater = setInterval(() => {
-      setCurrentTime(DateTime.now());
-      remainingLabMintues =
-        Interval.fromDateTimes(currentTime, eventInterval.end).count("minute") -
-        1;
-    }, 60000);
-
-    const percentage =
-      -1 *
-      Math.ceil(
-        remainingLabMintues / (eventInterval.count("minute") - 1) / 0.01 - 100
-      );
-
-    if (percentage == 100) clearInterval(timeUpdater);
-
-    return (
-      <div style={{ alignSelf: "end" }}>
-        <ProgressBar
-          completed={percentage}
-          bgColor="#52FF00"
-          borderRadius="0"
-          baseBgColor="#000000"
-          labelColor="#e809090"
-        />
-      </div>
-    );
-  };
 
   return (
     <Wrapper>
       <strong>All times are based off {currentTime.zoneName}</strong>
-      <TimeLineContainer>
-        {events
-          .filter(
-            events =>
-              events.startWeekDay == currentTime.setZone("UTC-5").weekday
-          )
-          .map(
-            (
-              {
-                title,
-                startWeekDay,
-                startHour,
-                startMinute,
-                endWeekDay,
-                endHour,
-                endMinute,
-                location
-              },
-              i
-            ) => {
-              const eventStart = ConvertCentralTime(
-                startWeekDay,
-                startHour,
-                startMinute
-              );
-              const eventEnd = ConvertCentralTime(
-                endWeekDay,
-                endHour,
-                endMinute
-              );
-              return (
-                <TimeStamp key={i}>
-                  {title} <br />
-                  {eventStart.toFormat("t")}
-                  {" - "}
-                  {eventEnd.toFormat("t")} <br />
-                  {location} Room
-                </TimeStamp>
-              );
-            }
-          )}
-      </TimeLineContainer>
-
-      <LabsContainer>
-        {[
-          "General Coworking",
-          "Beginner Dev",
-          "Web / App Dev",
-          "Game Dev",
-          "Landing Zone"
-        ].map((labTitle, i) => {
-          const gridArea = labTitle.split(" ")[0];
-          let currentEvent = events.filter(
-            event =>
-              event.location == labTitle &&
-              event.startWeekDay == currentTime.setZone("UTC-5").weekday
-          );
-
-          return currentEvent.length == 0 ? (
-            <Lab key={i} style={{ gridArea }}>
-              <Time>No sessions today</Time>
-              <Title>{labTitle}</Title>
-              <AlignedDropdown
-                toggleBtnText="Join"
-                dropdownItems={
-                  <>
-                    <a href="#">Primary</a>
-                    <a href="#">Room 1</a>
-                    <a href="#">Room 2</a>
-                  </>
-                }
-              />
-            </Lab>
-          ) : (
-            currentEvent.map(
-              ({
-                startWeekDay,
-                startHour,
-                startMinute,
-                endWeekDay,
-                endHour,
-                endMinute
-              }) => {
-                const eventStart = ConvertCentralTime(
-                  startWeekDay,
-                  startHour,
-                  startMinute
-                );
-                const eventEnd = ConvertCentralTime(
-                  endWeekDay,
-                  endHour,
-                  endMinute
-                );
-                const eventInterval = Interval.fromDateTimes(
-                  eventStart,
-                  eventEnd
-                );
-                const isLabActive =
-                  !eventInterval.isAfter(currentTime) &&
-                  !eventInterval.isBefore(currentTime);
-                const hasEnded = !eventInterval.isAfter(currentTime);
-                return (
-                  <Lab key={i} style={{ gridArea }}>
-                    <Time>
-                      {isLabActive
-                        ? "Happening now!"
-                        : hasEnded
-                        ? "All sessions have ended"
-                        : eventStart.toFormat("t")}
-                    </Time>
-                    <Title>{labTitle}</Title>
-                    <AlignedDropdown
-                      toggleBtnText="Join"
-                      dropdownItems={
-                        <>
-                          <a href="#">Primary</a>
-                          <a href="#">Room 1</a>
-                          <a href="#">Room 2</a>
-                        </>
-                      }
-                    />
-                    {isLabActive && handleProgression(eventInterval)}
-                  </Lab>
-                );
-              }
-            )
-          );
-        })}
-      </LabsContainer>
+      <TimeLine events={events} currentTime={currentTime} />
+      <Labs
+        events={events}
+        currentTime={currentTime}
+        currentTimeSetter={setCurrentTime}
+      />
     </Wrapper>
   );
 };
